+++ myapp/config/db.py
@@ -24,8 +24,9 @@ def close_db(e=None):
 def init_db():
     db = get_db()
 
-    with current_app.open_resource('schema.sql') as f:
+    with current_app.open_resource('config/schema.sql') as f:
         db.executescript(f.read().decode('utf8'))
+        print("O esquema do banco foi carregad com sucesso!")
 
 
 @click.command('init-db')
