+++ myapp/config/schema.sql
