+++ setup.cfg
@@ -5,7 +5,7 @@ url = https://flask.palletsprojects.com/tutorial/
 license = BSD-3-Clause
 maintainer = Pallets
 maintainer_email = contact@palletsprojects.com
-description = The basic blog app built in the Flask tutorial.
+description = Flask tutorial.
 long_description = file: README.rst
 long_description_content_type = text/x-rst
